feat: Manage multiple profiles
