+++ components/centraldashboard/public/components/manage-users-view_test.js
@@ -17,6 +17,7 @@ const TEMPLATE = `
 `;
 const user = 'test@kubeflow.org';
 const oNs = {namespace: 'ns1', role: 'owner'};
+const multiONs= [oNs];
 const generalNs = [oNs, {namespace: 'ns2', role: 'contributor', user}, {namespace: 'ns3', role: 'contributor', user}];
 
 describe('Manage Users View', () => {
@@ -54,14 +55,8 @@ describe('Manage Users View', () => {
     });
 
     it('UI State should show user / contribs and unhide sections when namespace available', async () => {
-        const contribList = ['foo@kubeflow.org', 'bar@kubeflow.org'];
-        mockIronAjax(
-            manageUsersView.$.GetContribsAjax,
-            contribList,
-        );
-
         manageUsersView.user = user;
-        manageUsersView.ownedNamespace = oNs;
+        manageUsersView.multiOwnedNamespaces = multiONs;
         manageUsersView.namespaces = generalNs;
 
         flush();
@@ -74,20 +69,12 @@ describe('Manage Users View', () => {
         expect(manageUsersView.shadowRoot.querySelector('.Contributors')
             .hasAttribute('hidden')).toBe(false, 'Contributors was still hidden');
         expect(manageUsersView.shadowRoot.querySelector('.Cluster-Namespaces')
-            .hasAttribute('hidden')).toBe(true, 'Cluster Namespaces should have been hidden');
-        expect(manageUsersView.shadowRoot.querySelector('.Contributors > h2 > .text').innerText)
-            .toBe('Contributors to your namespace - ns1');
-
+            .hasAttribute('hidden')).toBe(true, 'Cluster Namespaces should have been hidden');       
         // View prop expectations
         expect(manageUsersView.shadowRoot.querySelector('.Namespaces vaadin-grid').items)
             .toEqual(
-                [['ns1', 'Owner'], ['ns2, ns3', 'Contributor']],
+                [['ns1', 'Owner'], ['ns2', 'Contributor'], ['ns3', 'Contributor']],
                 'Invalid namespace memberships'
-            );
-        expect(manageUsersView.contributorList)
-            .toEqual(
-                contribList,
-                'Invalid list of contributors'
             );
     });
 
@@ -97,17 +84,13 @@ describe('Manage Users View', () => {
             ['ns1', user, contribList.join(', ')],
         ];
         mockIronAjax(
-            manageUsersView.$.GetContribsAjax,
-            contribList,
-        );
-        mockIronAjax(
             manageUsersView.$.GetAllNamespacesAjax,
             allPeeps,
         );
 
         manageUsersView.user = user;
         manageUsersView.isClusterAdmin = true;
-        manageUsersView.ownedNamespace = oNs;
+        manageUsersView.multiOwnedNamespaces = multiONs;
         manageUsersView.namespaces = [oNs];
 
         flush();
@@ -121,92 +104,6 @@ describe('Manage Users View', () => {
             .toEqual(
                 allPeeps,
                 'Invalid list of all namespace memberships'
-            );
-    });
-
-    it('Should handle errors correctly', async () => {
-        mockIronAjax(
-            manageUsersView.$.GetContribsAjax,
-            'Failed for test',
-            true,
-        );
-
-        manageUsersView.user = user;
-        manageUsersView.ownedNamespace = oNs;
-        manageUsersView.namespaces = [oNs];
-
-        flush();
-        await yieldForRequests();
-
-        expect(manageUsersView.$.ContribError.opened)
-            .toBe(
-                true,
-                'Error toast is not opened'
-            );
-        expect(manageUsersView.contribError)
-            .toBe('Failed for test');
-    });
-
-    it('Should add contributors correctly', async () => {
-        const contribList = ['foo@kubeflow.org', 'bar@kubeflow.org'];
-        const verificationContribs = ['ap@kubeflow.org'];
-        mockIronAjax(
-            manageUsersView.$.GetContribsAjax,
-            contribList,
-        );
-        mockIronAjax(
-            manageUsersView.$.AddContribAjax,
-            verificationContribs,
-        );
-
-        manageUsersView.user = user;
-        manageUsersView.ownedNamespace = oNs;
-        manageUsersView.namespaces = generalNs;
-
-        flush();
-        await yieldForRequests();
-
-        const input = manageUsersView.shadowRoot.querySelector('.Contributors md2-input');
-        input.value = 'new@google.com';
-        input.fireEnter();
-
-        await yieldForRequests();
-
-        expect(manageUsersView.contributorList)
-            .toEqual(
-                verificationContribs,
-                'Invalid list of contributors'
-            );
-    });
-
-    it('Should remove contributors correctly', async () => {
-        const contribList = ['foo@kubeflow.org', 'bar@kubeflow.org'];
-        const verificationContribs = ['ap@kubeflow.org'];
-        mockIronAjax(
-            manageUsersView.$.GetContribsAjax,
-            contribList,
-        );
-        mockIronAjax(
-            manageUsersView.$.RemoveContribAjax,
-            verificationContribs,
-        );
-
-        manageUsersView.user = user;
-        manageUsersView.ownedNamespace = oNs;
-        manageUsersView.namespaces = generalNs;
-
-        flush();
-        await yieldForRequests();
-
-        const chip = manageUsersView.shadowRoot.querySelector('.Contributors md2-input paper-chip:nth-of-type(1)');
-        chip.fireRemove({});
-
-        await yieldForRequests();
-
-        expect(manageUsersView.contributorList)
-            .toEqual(
-                verificationContribs,
-                'Invalid list of contributors'
             );
     });
 });
